Move model files and update import statements
