--- src/app/models/application.js
@@ -1,7 +1,0 @@
-// import Backbone from 'backbone';
-//
-//
-//
-//
-//
-// export default Application;
